Bump Mailer module to 3.1, fix Docker cron issues
Cron issues were linked to the use of non-root user, which was
running cron in the foreground. The user didn't have sufficient
access rights - root needs to run "cron -f".

On the other hand, the actual command in the crontab should be
run as non-root user. However non-root user cannot send output
of the command to the random process (PID 1) file descriptor.
Because of that, cron's task is executed as "root", which then
executes runs the command as non-root user via "runuser".
This way we can run our tasks as non-root user and forward
output to the other PID file descriptors.